PIC FW
This is the PIC16F877's firmware project folder. I used some old MPLAB X to have my Pickit 2 programmer supported. Some interesting filenames are there, but whatever, I was annoyed. The code contains a Knight Rider Chaser light, and a Serial UART connection to the nano. See the code as well for the commands the PIC understands. There is an error somewhere here, or in the Nano's code that during receiving messages from the PIC, some chars are garbled, but it's otherwise not much of an issue, for my use case at least.